Refactor into pending_calibration.py
